@fix: comprehensive code audit - fix 114 issues across full stack
Systematic code audit covering backend (Java), frontend (Vue), and mobile (KMP).

Critical fixes (7):
- Backend: fix unauthorized wallet recharge, card API IDOR, order stats leak, TOTP recovery logic
- Mobile: fix certificate pinning placeholder, logout token order, getUserCards IDOR

High fixes (27):
- Backend: card verification logic, file upload security, HTML sanitization, JWT cache, refresh token validation
- Frontend: OAuth token URL leak, localStorage role tampering
- Mobile: HTTP URL bypass, token refresh logic, machine code uniqueness, email bombing

Medium fixes (54):
- Backend: Swagger auth, SSRF protection, JPA entities, batch insert, pagination, OrderStatus enum
- Frontend: DOMPurify XSS protection, ARIA accessibility, canvas resize, code dedup
- Mobile: thread safety, ProGuard rules, SDK 35, ErrorMapper, atomic file write

Low fixes (26):
- Code quality, comments, magic numbers, logging, dependency cleanup

Test results: backend mvn test 59/59 passed, frontend npm build success, mobile gradle compile success.
Audit report: docs/plans/2026-06-20-code-audit-report.md

+++ backend/src/main/java/org/xxg/backend/backend/controller/ApiKeyController.java
@@ -121,20 +121,20 @@ public class ApiKeyController {
     @PostMapping("/{id}/users")
     @Transactional
     public ResponseEntity<ApiResponse<Void>> assignUser(@PathVariable Integer id, @RequestBody Map<String, Object> body) {
-        Long userId = body.get("userId") != null ? Long.valueOf(body.get("userId").toString()) : null;
+        Integer userId = body.get("userId") != null ? Integer.valueOf(body.get("userId").toString()) : null;
         if (userId == null) {
             throw new BusinessException("用户ID不能为空");
         }
         // 检查 API Key 是否存在
         apiKeyService.getApiKeyById(id);
         // 检查是否已分配
-        Optional<UserApiKey> existing = userApiKeyRepository.findByUserIdAndApiKeyId(userId, id.longValue());
+        Optional<UserApiKey> existing = userApiKeyRepository.findByUserIdAndApiKeyId(userId, id);
         if (existing.isPresent()) {
             throw new BusinessException("该用户已分配此 API Key");
         }
         UserApiKey userApiKey = new UserApiKey();
         userApiKey.setUserId(userId);
-        userApiKey.setApiKeyId(id.longValue());
+        userApiKey.setApiKeyId(id);
         userApiKey.setAssignTime(LocalDateTime.now());
         userApiKeyRepository.save(userApiKey);
         return ResponseEntity.ok(ApiResponse.ok("用户分配成功"));
@@ -151,7 +151,7 @@ public class ApiKeyController {
     @DeleteMapping("/{id}/users/{userId}")
     @Transactional
     public ResponseEntity<ApiResponse<Void>> unassignUser(@PathVariable Integer id, @PathVariable Integer userId) {
-        userApiKeyRepository.findByUserIdAndApiKeyId(userId.longValue(), id.longValue())
+        userApiKeyRepository.findByUserIdAndApiKeyId(userId, id)
                 .ifPresent(userApiKeyRepository::delete);
         return ResponseEntity.ok(ApiResponse.ok("用户取消分配成功"));
     }
